fetch: Disable Referer header.
This appears to cause problems when downloading from SourceForge, and
unfortunately there are still a lot of packages that do this, despite it
being a phishing cesspool.

On the flip side, this required re-using a single client instance rather
than each thread having its own, and that appears to improve performance
quite a bit!