fetch: Go back to printing initial status bar.
While debugging the SourceForge issue, it was more annoying not seeing
any output at all than the full status bar.  Perhaps at some point we
can add support for showing that we're in the process of following
redirects (which is why it takes so long), but for now this helps show
that we're at least in the fetch target.