check-shlibs: Add initial implementation.
As far as I have tested this is fully compatible with the awk versions,
with some additional features.  For example Mach-O libraries now support
CHECK_WRKREF_EXTRA_DIRS and CHECK_SHLIBS_TOXIC.

This implementation is, as expected, quite a bit faster.  Running the
_check-shlibs target in x11/kde-workspace4 improves the runtime from:

  real    0m20.441s
  user    0m7.540s
  sys     0m11.222s

to just:

  real    0m0.607s
  user    0m0.400s
  sys     0m0.242s

Tested in various scenarios on macOS and SmartOS.