check-portability: Add initial implementation.
This is mostly compatible with check-portability.sh, but does not yet
support:

 * CHECK_PORTABILITY_EXPERIMENTAL.

 * /bin/sh warning when running on Solaris 10 or older.

 * The weird $patched_files thing.

It also does not hardcode a skiplist of files, so ends up reading the
first line from more files than the shell script does.  However it is
still significantly faster.  Running in x11/qt5-qtwebengine goes from:

  real    0m52.145s
  user    0m43.979s
  sys     0m10.741s

to just:

  real    0m3.302s
  user    0m0.831s
  sys     0m2.469s

It remains to be seen whether the missing features are worth
implementing.